+++ EssentialFeed/EssentialFeediOSTests/Helpers/XCTestCase+Snapshot.swift
@@ -1,0 +1,8 @@
+//
+//  XCTestCase+Snapshot.swift.swift
+//  EssentialFeediOSTests
+//
+//  Created by Prabhat Tiwari on 31/08/25.
+//
+
+import Foundation
